pyproject.toml
[project]
name = "oceanbase_mcp_server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server that enables secure interaction with OceanBase databases. This server allows AI assistants to list tables, read data, and execute SQL queries through a controlled interface, making database exploration and analysis safer and more structured."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "mysql-connector-python>=9.1.0",
]
[[project.authors]]
name = "yaunoOo"
email = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
oceanbase_mcp_server = "oceanbase_mcp_server:main"